requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.20.0
sqlalchemy==1.4.46
alembic==1.8.1
pydantic==1.10.13
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pyjwt==2.6.0
requests==2.28.1
cryptography==39.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0